Remove svp fund tx logic from federation storage provider, and add it to bridge storage provider
