+++ rskj-core/src/main/java/co/rsk/peg/federation/FederationStorageProvider.java
@@ -1,7 +1,6 @@
 package co.rsk.peg.federation;
 
 import co.rsk.bitcoinj.core.NetworkParameters;
-import co.rsk.bitcoinj.core.Sha256Hash;
 import co.rsk.bitcoinj.core.UTXO;
 import co.rsk.bitcoinj.script.Script;
 import co.rsk.peg.federation.constants.FederationConstants;
@@ -40,10 +39,6 @@ public interface FederationStorageProvider {
 
     Optional<Script> getLastRetiredFederationP2SHScript(ActivationConfig.ForBlock activations);
     void setLastRetiredFederationP2SHScript(Script lastRetiredFederationP2SHScript);
-
-    void setSvpFundTxHashUnsigned(Sha256Hash hash);
-
-    Optional<Sha256Hash> getSvpFundTxHashUnsigned(ActivationConfig.ForBlock activations);
 
     void save(NetworkParameters networkParameters, ActivationConfig.ForBlock activations);
 }
